Rename ipc -> unix_socket_server
