+++ src/lib.rs
@@ -6,7 +6,6 @@ pub mod client;
 pub mod command;
 pub mod constants;
 pub mod event;
-pub mod ipc;
 pub mod ledger;
 pub mod mina_blocks;
 pub mod proof_systems;
@@ -16,4 +15,5 @@ pub mod server;
 pub mod snark_work;
 pub mod state;
 pub mod store;
+pub mod unix_socket_server;
 pub mod web;
